Prepare entity and collection for admin usage
