+++ src/data/index.ts
@@ -1,8 +1,10 @@
-import Entity from './Entity';
-import EntityCollection from './EntityCollection';
+import Criteria from './_internals/Criteria';
+import Entity from './_internals/Entity';
+import EntityCollection from './_internals/EntityCollection';
 
 export default {
   Classes: {
+    Criteria,
     Entity,
     EntityCollection,
   },
